requirements.txt
langchain
langgraph
langchain-openai
langchain-community
faiss-cpu
tiktoken
pydantic
redis
python-dotenv